+++ pushnotifications/src/androidTest/java/com/pusher/pushnotifications/ServerSyncProcessHandlerTest.kt
@@ -4,6 +4,7 @@ import android.support.test.InstrumentationRegistry
 import android.support.test.runner.AndroidJUnit4
 import com.pusher.pushnotifications.api.DeviceMetadata
 import com.pusher.pushnotifications.api.PushNotificationsAPI
+import com.pusher.pushnotifications.api.PushNotificationsAPIUnprocessableEntity
 import com.pusher.pushnotifications.auth.TokenProvider
 import com.pusher.pushnotifications.internal.*
 import com.squareup.moshi.Moshi
@@ -661,6 +662,40 @@ class ServerSyncProcessHandlerTest {
     assertThat((lastHandleServerSyncEvent!! as UserIdSet).userId, `is`(equalTo(userId)))
     assertNotNull((lastHandleServerSyncEvent!! as UserIdSet).pusherCallbackError)
     assertThat((lastHandleServerSyncEvent!! as UserIdSet).pusherCallbackError!!.cause, `is`(equalTo(ExceptionalTokenProvider.exception)))
+  }
+
+  @Test
+  fun setUserIdShouldReturnErrorEventIfTheUserHasTooManyDevicesAlready() {
+    val userId = "alice"
+    tokenProvider = StubTokenProvider()
+
+    assertThat(mockServer.requestCount, `is`(equalTo(0)))
+    assertNull(deviceStateStore.userId)
+
+    val startJob = ServerSyncHandler.start("token-123", emptyList())
+    jobQueue.push(startJob.obj as ServerSyncJob)
+
+    // expect register device
+    mockServer.enqueue(MockResponse().setBody("""{"id": "d-123", "initialInterestSet": []}"""))
+
+    handler.handleMessage(startJob)
+
+    assertThat(mockServer.requestCount, `is`(equalTo(1)))
+    assertNotNull(deviceStateStore.deviceId)
+    assertNull(deviceStateStore.userId)
+
+    val setUserIdJob = ServerSyncHandler.setUserId(userId)
+    jobQueue.push(setUserIdJob.obj as ServerSyncJob)
+
+    // expect set user id returning a 422 unprocessable entity
+    mockServer.enqueue(MockResponse().setBody("{}").setResponseCode(422))
+
+    handler.handleMessage(setUserIdJob)
+
+    assertNotNull(lastHandleServerSyncEvent)
+    assertThat((lastHandleServerSyncEvent!! as UserIdSet).userId, `is`(equalTo(userId)))
+    assertNotNull((lastHandleServerSyncEvent!! as UserIdSet).pusherCallbackError)
+    assertNotNull((lastHandleServerSyncEvent!! as UserIdSet).pusherCallbackError!!.cause is PushNotificationsAPIUnprocessableEntity)
   }
 
   @Test